clean verification
