+++ src/StateService/Verification/VerificationContext.cs
@@ -1,83 +1,142 @@
 using Akka.Actor;
-using Akka.Util.Internal;
+using Neo.Cryptography;
+using Neo.Cryptography.ECC;
+using Neo.IO;
+using Neo.Ledger;
+using Neo.Network.P2P;
 using Neo.Network.P2P.Payloads;
+using Neo.Persistence;
+using Neo.Plugins.StateService.Storage;
+using Neo.SmartContract;
+using Neo.SmartContract.Native;
 using Neo.Wallets;
-using System.Collections.Concurrent;
-using System.Linq;
+using System;
+using System.Collections.Generic;
 
 namespace Neo.Plugins.StateService.Verification
 {
     public class VerificationContext
     {
-        private const int MaxCachedVerificationProcessCount = 10;
+        private StateRoot root;
+        private ExtensiblePayload payload;
         private readonly Wallet wallet;
-        private readonly ConcurrentDictionary<uint, VerificationProcess> processes = new ConcurrentDictionary<uint, VerificationProcess>();
+        private KeyPair keyPair;
+        private int myIndex;
+        private uint rootIndex;
+        private ECPoint[] verifiers;
+        private int M => verifiers.Length - (verifiers.Length - 1) / 3;
+        private readonly Dictionary<int, byte[]> signatures = new Dictionary<int, byte[]>();
 
-        public VerificationContext(Wallet wallet)
+        public bool IsValidator => myIndex >= 0;
+        public int MyIndex => myIndex;
+        public ICancelable Timer;
+        public StateRoot StateRoot
+        {
+            get
+            {
+                if (root is null)
+                {
+                    using var snapshot = StateStore.Singleton.GetSnapshot();
+                    root = snapshot.GetStateRoot(rootIndex);
+                }
+                return root;
+            }
+        }
+        public ExtensiblePayload Message => payload;
+
+        public VerificationContext() { }
+
+        public VerificationContext(Wallet wallet, uint index)
         {
             this.wallet = wallet;
+            Initialize(index);
         }
 
-        public VerificationProcess NewProcess(uint index)
+        public void Initialize(uint index)
         {
-            if (MaxCachedVerificationProcessCount <= processes.Count)
+            myIndex = -1;
+            root = null;
+            rootIndex = index;
+            using SnapshotView snapshot = Blockchain.Singleton.GetSnapshot();
+            verifiers = NativeContract.RoleManagement.GetDesignatedByRole(snapshot, Role.StateValidator, index);
+            for (int i = 0; i < verifiers.Length; i++)
             {
-                var indexes = processes.Keys.OrderBy(i => i).ToArray();
-                while (MaxCachedVerificationProcessCount <= indexes.Length)
-                {
-                    if (processes.TryRemove(indexes[0], out var value))
-                    {
-                        value.Timer.CancelIfNotNull();
-                        indexes = indexes[..1];
-                    }
-                }
+                WalletAccount account = wallet?.GetAccount(verifiers[i]);
+                if (account?.HasKey != true) continue;
+                myIndex = i;
+                keyPair = account.GetKey();
+                break;
             }
-            var p = new VerificationProcess(wallet, index);
-            if (!p.IsValidator) return null;
-            if (processes.TryAdd(index, p))
-            {
-                Utility.Log(nameof(VerificationContext), LogLevel.Info, $"new validate process, height={index}, index={p.MyIndex}, ongoing={processes.Count}");
-                return p;
-            }
-            return null;
         }
 
-        public Vote CreateVote(uint index)
+        public Vote CreateVote()
         {
-            if (processes.TryGetValue(index, out VerificationProcess p))
+            if (StateRoot is null) return null;
+            if (!signatures.TryGetValue(myIndex, out byte[] sig))
             {
-                return p.CreateVote();
+                sig = StateRoot.Sign(keyPair);
+                signatures[myIndex] = sig;
             }
-            return null;
+            return new Vote(rootIndex, myIndex, sig);
         }
 
-        public bool OnVote(Vote vote)
+        public bool AddSignature(int index, byte[] sig)
         {
-            if (processes.TryGetValue(vote.RootIndex, out VerificationProcess p))
+            if (M <= signatures.Count) return false;
+            if (index < 0 || verifiers.Length <= index) return false;
+            if (signatures.ContainsKey(index)) return false;
+            Utility.Log(nameof(VerificationContext), LogLevel.Info, $"vote received, index={index}");
+            ECPoint validator = verifiers[index];
+            byte[] hash_data = StateRoot?.GetHashData();
+            if (hash_data is null || !Crypto.VerifySignature(hash_data, sig, validator))
             {
-                return p.AddSignature(vote.ValidatorIndex, vote.Signature);
+                Utility.Log(nameof(VerificationContext), LogLevel.Info, "incorrect vote, invalid signature");
+                return false;
             }
-            return false;
+            signatures.Add(index, sig);
+            return true;
         }
 
-        public ExtensiblePayload CheckVotes(uint index)
+        private void CreateExtensiblePayload()
         {
-            if (processes.TryGetValue(index, out VerificationProcess p) && p.CheckSignatures())
+            Contract contract = Contract.CreateMultiSigContract(M, verifiers);
+            ContractParametersContext sc = new ContractParametersContext(StateRoot);
+            for (int i = 0, j = 0; i < verifiers.Length && j < M; i++)
             {
-                return p.Message;
+                bool ok = signatures.TryGetValue(i, out byte[] sig);
+                if (!ok) continue;
+                sc.AddSignature(contract, verifiers[i], sig);
+                j++;
             }
-            return null;
+            StateRoot.Witness = sc.GetWitnesses()[0];
+
+            payload = new ExtensiblePayload
+            {
+                Category = StatePlugin.StatePayloadCategory,
+                ValidBlockStart = StateRoot.Index,
+                ValidBlockEnd = uint.MaxValue,
+                Sender = keyPair.PublicKeyHash,
+                Data = StateRoot.ToArray(),
+            };
+            try
+            {
+                sc = new ContractParametersContext(payload);
+                wallet.Sign(sc);
+            }
+            catch (InvalidOperationException)
+            {
+                return;
+            }
+            payload.Witness = sc.GetWitnesses()[0];
         }
 
-        public void StopProcess(uint index)
+        public bool CheckSignatures()
         {
-            processes.Where(i => i.Key <= index).ForEach(i =>
-            {
-                if (processes.TryRemove(i.Key, out var value))
-                {
-                    value.Timer.CancelIfNotNull();
-                }
-            });
+            if (StateRoot is null) return false;
+            if (signatures.Count < M) return false;
+            if (StateRoot.Witness != null) return true;
+            CreateExtensiblePayload();
+            return true;
         }
     }
 }
